+++ backend/app/services/internal/auth_service.py
@@ -19,33 +19,6 @@ from app.core.exceptions.http_exception import UnauthorizedException, NotFoundEx
 from datetime import datetime, timedelta
 
 
-def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
-    """Create a JWT access token."""
-    to_encode = data.copy()
-    if expires_delta:
-        expire = datetime.utcnow() + expires_delta
-    else:
-        expire = datetime.utcnow() + timedelta(
-            minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES
-        )
-
-    to_encode.update({"exp": expire})
-
-    return encrypt_token(to_encode)
-
-
-def verify_token(token: str) -> Optional[str]:
-    """Verify and decode a JWT token."""
-    try:
-        payload = decrypt_token(token)
-        username: str = payload.get("sub")
-        if username is None:
-            return None
-        return username
-    except JWTError:
-        return None
-
-
 class AuthService:
 
     def __init__(self):
@@ -82,7 +55,7 @@ class AuthService:
         if not verify_password(login_data.password, user_data.password_hash):
             raise UnauthorizedException(message="Invalid email or password")
         self.user_repo.update_last_login(user_data.id)
-        access_token = create_access_token(data={"sub": user_data.email})
+        access_token = self.create_access_token(data={"sub": user_data.email})
         user_response = UserResponse.model_validate(user_data)
         return {"user": user_response.model_dump(mode='json'), "access_token": access_token}
 
@@ -96,3 +69,29 @@ class AuthService:
         new_password_hash = get_password_hash(new_password)
         updated_user = self.user_repo.update_password(user.id, new_password_hash)
         return {"message": "Password updated successfully"}
+
+    def create_access_token(self, data: dict, expires_delta: Optional[timedelta] = None):
+        """Create a JWT access token."""
+        to_encode = data.copy()
+        if expires_delta:
+            expire = datetime.utcnow() + expires_delta
+        else:
+            expire = datetime.utcnow() + timedelta(
+                minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES
+            )
+
+        to_encode.update({"exp": expire})
+
+        return encrypt_token(to_encode)
+
+
+    def verify_token(self, token: str) -> Optional[str]:
+        """Verify and decode a JWT token."""
+        try:
+            payload = decrypt_token(token)
+            username: str = payload.get("sub")
+            if username is None:
+                return None
+            return username
+        except JWTError:
+            return None
