+++ backend/app/utils/response_handler.py
@@ -1,4 +1,6 @@
 from typing import Any, Optional
+from fastapi import Request
+from fastapi.exceptions import RequestValidationError
 from fastapi.responses import JSONResponse
 
 class ResponseHandler:
@@ -25,3 +27,27 @@ class ResponseHandler:
                 "detail": errors
             },
         )
+
+
+async def validation_exception_handler(
+    request: Request,
+    exc: RequestValidationError
+):
+    return ResponseHandler.error_response(
+        message="Validation Error",
+        code=422,
+        errors={
+            "details": exc.errors(),
+            "path": request.url.path,
+        },
+    )
+
+
+async def internal_exception_handler(request: Request, exc: Exception):
+    return ResponseHandler.error_response(
+        message="Internal Server Error",
+        code=500,
+        errors={
+            "path": request.url.path
+        },
+    )
